refactor: relocate query handlers, validators to `Domain`, add `History` entity and update imports
- Moved `GetTournamentQueryHandler`, `GetUserQueryValidator`, and related logic to `Domain` layer.
- Added `History` entity for managing tournament history details.
- Updated import paths across affected modules for consistency.
