+++ tournament/src/Application/Services/Concrete/GetTournamentService.ts
@@ -3,10 +3,10 @@ import {FastifyReply} from "fastify";
 import {Result} from "../../../Shared/Utils/Result.js";
 import {Prisma} from '@prisma/client';
 import {GetTournamentDTO} from "../../DTO/ToQuery/GetTournamentDTO";
-import {GetTournamentQuery} from "../../Queries/QueryObject/GetTournamentQuery";
+import {GetTournamentQuery} from "../../../Domain/Queries/QueryObject/GetTournamentQuery";
 import {GetTournamentViewModel} from "../../ViewModel/GetTournamentViewModel";
 import {TournamentRepository} from "../../../Infrastructure/Persistence/Repositories/Concrete/TournamentRepository";
-import {GetTournamentQueryHandler} from "../../Queries/Handlers/GetTournamentQueryHandler";
+import {GetTournamentQueryHandler} from "../../../Domain/Queries/Handlers/GetTournamentQueryHandler";
 import {ErrorCatalog} from "../../../Shared/Errors/ErrorCatalog";
 import {ValidationException} from "../../../Shared/Errors/ValidationException";
 import {NotificationError} from "../../../Shared/Errors/NotificationError";
